+++ src/97saveas.js
@@ -1,0 +1,1 @@
+if((typeof exports != 'object') && (typeof importScripts != 'function')) {