+++ src/98saveas.js
@@ -1,1 +1,0 @@
-if(typeof exports != 'object') {