+++ test/test200.js
@@ -5,12 +5,13 @@ if(typeof exports === 'object') {
 	__dirname = '.';
 };
 
-describe('Test 200 IS NULL', function() {
+describe('Test 200 IS NULL + IS NOT NULL', function() {
 
     it("1. Simple Variant", function(done) {
-    	alasql('DROP TABLE IF EXISTS one');
+    	alasql('CREATE DATABASE test200; USE test200');
         var res = alasql('IF 1 IS NOT NULL CREATE TABLE one');
         assert(!!alasql.tables.one);
+        alasql('DROP DATABASE test200');
     	done();
     });